Assistant checkpoint: Fix blog editor data population issue
Assistant generated file changes:
- client/src/pages/blog-editor.tsx: Fix blog editor data population, Update TinyMCE editor initialization

---

User prompt:

Ok this is frustrating.  using the curl command we are able to see the api is retrieving the correct blog by id.  That data is still not populating in the blog editor.   We know the connection is allowing the titles to be populated in the /blog page, but not the editor.  Lets make sure the fields in the editor match to the associated data in the table and are able to retrieve and display it correctly for users to edit.